mwan3: comply with OpenWRT source based routing
OpenWRT uses source based routing which has filters on the outgoing
routes. These should be copied to the mwan3 routing tables so that
packets with a src IP specified are not routed out of an incorrect
interface.

The old behavior was potentially useful when mwan3track packets were
receiving the fwmark, but is no longer needed and could problematic,
as it allows packets with src_ip of one interface to be routed out of
another.

Signed-off-by: Aaron Goodman <aaronjg@stanford.edu>