mwan3: reload all routes on ifup events
on router startup, it is possible that the mwan3 routing tables are
created when an interface is in the process of being brought up. When
this happens, the upcoming interface can have routes in the "linkdown"
state, which are not added to the mwan3 routing tables. So we avoid
this by reloading all the routes when an interface comes online.

Signed-off-by: Aaron Goodman <aaronjg@stanford.edu>